requirements.txt
fastapi
uvicorn
python-dotenv
passlib
SQLAlchemy
alembic
PyJWT
python-jose
pydantic
fastapi[standard]
bcrypt